Final version (with correct naming convention)
Updates in this version: 

General Updates:
Changed naming to fit the convention: "ATOC4500_project_YourLastName_final.ipynb". 
Fixed small bug with fontsize arguments in plots, changed a typo on plot labels from m^2 to m^3, added detail to the descriptions of statistical methods, and edited the print statements on the Fisher-Z transformation for clarity. Improved the summary section by describing how the confidence intervals on the correlation relate to my hypothesis. 

Updated code for calculating statistics: 
Added code to standardize data, then re-calculated autocorrelation and effective sample size values. Corrected the T-test to be positive even if there is a negative correlation by taking the absolute value of r. 

Thank you to Jonah and Fergus for detailed feedback on the initial version of this code.